third_party/blink/web_tests/external/wpt/navigation-api/ordering-and-transition/anchor-download-intercept-reject_currententrychange-expected.txt

This is a testharness.js-based test.
[FAIL] event and promise ordering for <a download> intercepted by passing a rejected promise to intercept()
  assert_array_equals: expected property 3 to be "AbortSignal abort" but got "promise microtask" (expected array ["navigate", "currententrychange", "handler run", "AbortSignal abort", "navigateerror", "promise microtask", "transition.finished rejected"] got ["navigate", "currententrychange", "handler run", "promise microtask", "AbortSignal abort", "navigateerror", "transition.finished rejected"])
Harness: the test ran to completion.
